backend/NutritionalFunc/requirements.txt
azure-storage-blob
pandas
numpy
scikit-learn
flask
pyodbc
bcrypt
pyjwt
python-dotenv
requests